refactor: move game to backend, frontend
