+++ backend/core/asgi.py
@@ -8,9 +8,13 @@ https://docs.djangoproject.com/en/5.0/howto/deployment/asgi/
 """
 
 import os
+import threading
 
+from game.engine import start_server
 from django.core.asgi import get_asgi_application
 
 os.environ.setdefault('DJANGO_SETTINGS_MODULE', 'core.settings')
 
 application = get_asgi_application()
+
+threading.Thread(target=start_server, args=(4444,)).start() # TODO: game port move to env
