+++ backend/requirements.txt
@@ -1,9 +1,29 @@
 asgiref==3.7.2
+attrs==23.2.0
+autobahn==23.6.2
+Automat==22.10.0
+cffi==1.16.0
+constantly==23.10.4
+cryptography==41.0.7
+daphne==4.0.0
 Django==5.0
 django-cors-headers==4.3.1
 djangorestframework==3.14.0
+hyperlink==21.0.0
+idna==3.6
+incremental==22.10.0
 psycopg==3.1.16
 psycopg2-binary==2.9.9
+pyasn1==0.5.1
+pyasn1-modules==0.3.0
+pycparser==2.21
+pyOpenSSL==23.3.0
 pytz==2023.3.post1
+service-identity==23.1.0
+six==1.16.0
 sqlparse==0.4.4
+Twisted==23.10.0
+txaio==23.1.1
 typing_extensions==4.9.0
+websockets==12.0
+zope.interface==6.1
