+++ frontend/public/game/phong.html
@@ -10,6 +10,6 @@
     </script> -->
   </head>
   <body>
-    <script type="module" src="/static/main.js"></script>
+    <script type="module" src="/src/game/phong.js"></script>
   </body>
 </html>
