--- game/.dockerignore
@@ -1,2 +1,0 @@
-Dockerfile
-.dockerignore
