--- game/core/asgi.py
@@ -1,20 +1,0 @@
-"""
-ASGI config for core project.
-
-It exposes the ASGI callable as a module-level variable named ``application``.
-
-For more information on this file, see
-https://docs.djangoproject.com/en/5.0/howto/deployment/asgi/
-"""
-
-import os
-import threading
-
-from engine import engine
-from django.core.asgi import get_asgi_application
-
-os.environ.setdefault('DJANGO_SETTINGS_MODULE', 'core.settings')
-
-application = get_asgi_application()
-
-threading.Thread(target=engine.start_server, args=(5555,)).start() # TODO: game port move to env
