--- game/core/wsgi.py
@@ -1,20 +1,0 @@
-"""
-WSGI config for core project.
-
-It exposes the WSGI callable as a module-level variable named ``application``.
-
-For more information on this file, see
-https://docs.djangoproject.com/en/5.0/howto/deployment/wsgi/
-"""
-
-import os
-import threading
-
-from engine import engine
-from django.core.wsgi import get_wsgi_application
-
-os.environ.setdefault('DJANGO_SETTINGS_MODULE', 'core.settings')
-
-application = get_wsgi_application()
-
-threading.Thread(target=engine.start_server, args=(5555,)).start() # TODO: game port move to env
