--- game/static/client.js
@@ -1,3 +1,0 @@
-export default {
-    // socket,
-};
